Test/Test_System/requirements.txt
python-can==4.4.2
PyYAML==6.0.2
Jinja2==3.1.4